Relocate the deploy config into the properties
