+++ java/clients/src/main/java/sleeper/clients/deploy/DeployNewInstance.java
@@ -35,9 +35,9 @@ import sleeper.clients.util.CommandPipelineRunner;
 import sleeper.clients.util.EcrRepositoryCreator;
 import sleeper.clients.util.cdk.CdkCommand;
 import sleeper.clients.util.cdk.InvokeCdkForInstance;
-import sleeper.configuration.deploy.DeployInstanceConfiguration;
-import sleeper.configuration.deploy.DeployInstanceConfigurationFromTemplates;
 import sleeper.configuration.properties.SleeperPropertiesValidationReporter;
+import sleeper.configuration.properties.deploy.DeployInstanceConfiguration;
+import sleeper.configuration.properties.deploy.DeployInstanceConfigurationFromTemplates;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.local.SaveLocalProperties;
 import sleeper.configuration.properties.table.TableProperties;
