+++ java/clients/src/main/java/sleeper/clients/deploy/PopulateInstanceProperties.java
@@ -20,8 +20,8 @@ import com.amazonaws.services.securitytoken.AWSSecurityTokenService;
 import com.amazonaws.services.securitytoken.model.GetCallerIdentityRequest;
 import software.amazon.awssdk.regions.providers.AwsRegionProvider;
 
-import sleeper.configuration.deploy.DeployInstanceConfiguration;
 import sleeper.configuration.properties.SleeperScheduleRule;
+import sleeper.configuration.properties.deploy.DeployInstanceConfiguration;
 import sleeper.configuration.properties.instance.InstanceProperties;
 
 import java.nio.file.Path;
