+++ java/configuration/src/main/java/sleeper/configuration/properties/deploy/DeployInstanceConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.configuration.deploy;
+package sleeper.configuration.properties.deploy;
 
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
