+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/testutil/LocalStackTestInstance.java
@@ -15,7 +15,7 @@
  */
 package sleeper.systemtest.drivers.testutil;
 
-import sleeper.configuration.deploy.DeployInstanceConfiguration;
+import sleeper.configuration.properties.deploy.DeployInstanceConfiguration;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.schema.Field;
