+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/fixtures/SystemTestInstanceTest.java
@@ -18,7 +18,7 @@ package sleeper.systemtest.suite.fixtures;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.configuration.deploy.DeployInstanceConfiguration;
+import sleeper.configuration.properties.deploy.DeployInstanceConfiguration;
 import sleeper.configuration.properties.instance.CommonProperty;
 import sleeper.configuration.properties.table.TableProperty;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceConfiguration;
